requirements.txt
aiogram==3.15.0
psutil==6.1.1
python-dotenv==1.0.1
